refactor graphQL tests
